finished the function
Finished the onUpdateProfile function
This now triggers on the submit button on the form of the profile.

Gets the username from the .username element (this is populated on page load)

Calls the getUpdateProfileFormData function this returns the input values from the update profile form.

Calls the updateProfile function
By passing in name and the form data.  THis sends the put request to the api.

It then updates the ui with the newly created data by calling the readprofile function
and passing it into the updateProfileDetails function.

In short it updates the profile info with a put request and updates the profile ui to reflect those changes.